requirements.txt
faiss-cpu
numpy
openai